Allow PATCH on NodeResource, add unit tests for NodeResource, consolidate all unit tests for core APIs, and remove unused GuardianTastypieAuthz.py
